% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkRecode}
\alias{checkRecode}
\title{Check Recode List}
\usage{
checkRecode(recodeList, columnNames = NULL)
}
\arguments{
\item{recodeList}{A recode list.}

\item{columnNames}{Character Vector of columns names that should be checked. Defaults to c("newValues", "oldValues").}
}
\description{
Check Recode List
}
\examples{
# tbd
}
