% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRecodeList.R
\name{checkRecodeList}
\alias{checkRecodeList}
\title{Check Recode List}
\usage{
checkRecodeList(recodeList)
}
\arguments{
\item{recodeList}{A recode list.}
}
\description{
Check Recode List.
}
\examples{
# tbd

}
