% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRecodeDB.R
\name{createRecodeDB}
\alias{createRecodeDB}
\title{Create Recode Data Base}
\usage{
createRecodeDB(recodeListList, filePath)
}
\arguments{
\item{recodeListList}{A named list of \code{recodeLists}.}

\item{filePath}{Path to the \code{.xlsx} file in which the data base should be stored.}
}
\description{
Create a recode data base.
}
\examples{
# tbd

}
