% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractManualRecode.R
\name{extractManualRecode}
\alias{extractManualRecode}
\title{Get manual recodes}
\usage{
extractManualRecode(recodedList, varName)
}
\arguments{
\item{recodedList}{A data frame with the automatically recoded values.}

\item{varName}{Character string for the column containing the automatically recoded values.}
}
\value{
Returns a data frame which only includes the values that have to be recoded manually. It can be saved to Excel for manual recoding.
}
\description{
Get manual recodes
}
\examples{
# example data frame
df <- data.frame(id = 1:4,
                 country = c("Germany", "Scotland", NA , "Egypt"),
                 capital = c("Berlin", "Edinburgh", "Stockholm", "Cairo"))
# extract values to recode manually
manual_recodes <- extractManualRecode(recodedList = df, varName = "country")
print(manual_recodes)
# export to excel, edit, import
tmp <- writexl::write_xlsx(manual_recodes, path = "manual_recodes.xlsx")
readxl::read_xlsx(tmp)
file.remove(tmp)

}
