% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecodeDB.R
\name{getRecodeDB}
\alias{getRecodeDB}
\title{Get Recode Data Base}
\usage{
getRecodeDB(filePath)
}
\arguments{
\item{filePath}{Path to the \code{.xlsx} file in which the data base is stored.}
}
\value{
A recode list.
}
\description{
Get a complete recode data base.
}
\examples{
# tbd

}
