% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecodeDB.R
\name{getRecodeList}
\alias{getRecodeList}
\title{Get Recode List}
\usage{
getRecodeList(filePath, name)
}
\arguments{
\item{filePath}{Path to the \code{.xlsx} file in which the data base is stored.}

\item{name}{Name of the specific recode list.}
}
\value{
A recode list.
}
\description{
Get a recode list.
}
\examples{
# tbd

}
