% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualRecode.R
\name{manualRecode}
\alias{manualRecode}
\title{Get manual recodes}
\usage{
manualRecode(recodedList)
}
\arguments{
\item{recodedList}{A data frame with the automatically recoded values.}
}
\value{
Returns a data frame wich only includes the values that have to be recoded manually. This can be saved to excel for manual recoding.
}
\description{
Get manual recodes
}
\examples{
# tbd
}
