% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecodeDB.R
\name{namesRecodeDB}
\alias{namesRecodeDB}
\title{Get Names of Recode Data Base}
\usage{
namesRecodeDB(filePath)
}
\arguments{
\item{filePath}{Path to the \code{.xlsx} file in which the data base is stored.}
}
\value{
A recode list.
}
\description{
Get the names of the individual recode lists within a recode data base.
}
\examples{
# tbd

}
