% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRecodeDB.R
\name{updateRecodeDB}
\alias{updateRecodeDB}
\title{Update Recode Data Base}
\usage{
updateRecodeDB(newRecodes, recodeDBPath, name)
}
\arguments{
\item{newRecodes}{A \code{data.frame} containing new recode information.}

\item{recodeDBPath}{Path to the \code{.xlsx} file in which the data base is stored.}

\item{name}{Name of the specific recode list.}
}
\description{
Update a recode data base.
}
\examples{
# tbd

}
