% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkRecodeList}
\alias{checkRecodeList}
\title{Check Recode List}
\usage{
checkRecodeList(recodeList)
}
\arguments{
\item{recodeList}{A recode list: a data frame with the columns "oldValues" and "newValues".}
}
\description{
Checks whether a data frame called 'recodeList' has columns 'oldValues', containing the values to be recoded, and 'newValues', containing the new values which recode the old ones.
}
\examples{
recodeList <- data.frame(oldValues = c("Berlin", "Kairo", "Englant", "Schottland"),
                         newValues = c("Deutschland", "Ägypten", "England", "Schottland"))
checkRecodeList(recodeList)
}
