% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRecodeDB.R
\name{createRecodeDB}
\alias{createRecodeDB}
\title{Create Recode Data Base}
\usage{
createRecodeDB(recodeListList, filePath)
}
\arguments{
\item{recodeListList}{A named list of \code{recodeLists}.}

\item{filePath}{Path to the \code{.xlsx} file in which the data base should be stored.}
}
\description{
Create and store a recode data base in an Excel file.
The Excel file will have sheets corresponding to each recodeList name.
}
\examples{
# create a named list of data frames
recodeListList <- list( Europe = data.frame(
                                  id = 1:4,
                                  oldValues = c("Berlin", "Copenhagen", "Rome", "Madrid"),
                                  newValues = c("Germany", "Denmark", "Italy" , "Spain")),
                        Asia = data.frame(
                                  id = 1:4,
                                  oldValues = c("Baku", "Tokyo", "Kathmandu", "Singapore"),
                                  newValues = c("Azerbaijan", "Japan", "Nepal" , "Singapore")))
print(recodeListList)
filePath <- paste0(tempdir(),"\\\\recodeListList.xlsx")
# create the Excel-file
createRecodeDB(recodeListList = recodeListList, filePath = filePath)

}
