% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRecodeDB.R
\name{createRecodeDB}
\alias{createRecodeDB}
\title{Create Recode Data Base}
\usage{
createRecodeDB(recodeListList, filePath, fileType = "csv")
}
\arguments{
\item{recodeListList}{A named list of \code{recodeLists}.}

\item{filePath}{Path to the \code{.csv} or \code{.xlsx} file in which the data base should be stored.}

\item{fileType}{"csv2" (default), "csv", "xlsx"}
}
\description{
Create and store a recode data base in an Excel (\code{.csv} or \code{.xlsx}) file.
For each recodeList, a separate \code{.csv} file will be generated.
The \code{.xlsx} file will have sheets corresponding to each recodeList name.
}
\examples{
# create a named list of data frames
recodeListList <- list( Europe = data.frame(
                                  id = 1:4,
                                  oldValues = c("Berlin", "Copenhagen", "Rome", "Madrid"),
                                  newValues = c("Germany", "Denmark", "Italy" , "Spain")),
                        Asia = data.frame(
                                  id = 1:4,
                                  oldValues = c("Baku", "Tokyo", "Kathmandu", "Singapore"),
                                  newValues = c("Azerbaijan", "Japan", "Nepal" , "Singapore")))
recodeListList
# create .csv file with comma as decimal point (default)
filePath_temp <- tempfile(fileext = ".csv")
createRecodeDB(recodeListList = recodeListList, filePath = filePath_temp, fileType = "csv2")
createRecodeDB(recodeListList = recodeListList, filePath = filePath_temp)
# create .csv file
createRecodeDB(recodeListList = recodeListList, filePath = filePath_temp, fileType = "csv")
# create .xlsx file
filePath_temp <- tempfile(fileext = ".xlsx")
createRecodeDB(recodeListList = recodeListList, filePath = filePath_temp, fileType = "xlsx")

}
