% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractManualRecode.R
\name{extractManualRecode}
\alias{extractManualRecode}
\title{Get manual recodes}
\usage{
extractManualRecode(recodedList, varName)
}
\arguments{
\item{recodedList}{A data frame with the automatically recoded values.+}

\item{varName}{Character string for the column containing the automatically recoded values.}
}
\value{
Returns a data frame wich only includes the values that have to be recoded manually. This can be saved to excel for manual recoding.
}
\description{
Get manual recodes
}
\examples{
# tbd
}
