% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecodeDB.R
\name{getRecodeList}
\alias{getRecodeList}
\title{Get Recode List}
\usage{
getRecodeList(directory = getwd(), DBname, ListName, fileType = "csv2")
}
\arguments{
\item{directory}{Path to the directory where the data base is stored.}

\item{DBname}{Name of the database.}

\item{ListName}{Name of the specific recode list to be imported.}

\item{fileType}{"csv2" (default), "csv", "xlsx"}
}
\value{
A recode list.
}
\description{
Import a recode list from an Excel file containing a data base of different \code{recodeLists}.
}
\examples{
# Create recode list data base using `createRecodeDB`
Countries <- list( Europe = data.frame(
                                  id = 1:4,
                                  oldValues = c("Berlin", "Copenhagen", "Rome", "Madrid"),
                                  newValues = c("Germany", "Denmark", "Italy" , "Spain")),
                   Asia = data.frame(
                                  id = 1:4,
                                  oldValues = c("Baku", "Tokyo", "Kathmandu", "Singapore"),
                                  newValues = c("Azerbaijan", "Japan", "Nepal" , "Singapore")))
Countries
directory <- tempdir()
createRecodeDB(recodeListList = Countries,
               directory = directory,
               DBname = "Countries",
               fileType = "csv2")
createRecodeDB(recodeListList = Countries,
               directory = directory,
               DBname = "Countries",
               fileType = "xlsx")
# Import lists from csv data base
getRecodeList(directory = directory, DBname = "Countries", ListName = "Europe", fileType = "csv2")
getRecodeList(directory = directory, DBname = "Countries", ListName = "Asia", fileType = "csv2")
# Import lists from xlsx data base
getRecodeList(directory = directory, DBname = "Countries", ListName = "Europe", fileType = "xlsx")
getRecodeList(directory = directory, DBname = "Countries", ListName = "Asia", fileType = "xlsx")
}
