% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecodeDB.R
\name{namesRecodeDB}
\alias{namesRecodeDB}
\title{Get Names of Recode Data Base}
\usage{
namesRecodeDB(directory, DBname, fileType = "csv2")
}
\arguments{
\item{directory}{Path to the directory where the data base is stored.}

\item{DBname}{Name of the database.}

\item{fileType}{\code{csv2} (default), \code{csv}, \code{xlsx}}
}
\description{
Get the names of the individual recode lists within a recode data base.
}
\examples{
# Create recode list data base using `createRecodeDB`
Countries <- list( Europe = data.frame(
                                  id = 1:4,
                                  oldValues = c("Berlin", "Copenhagen", "Rome", "Madrid"),
                                  newValues = c("Germany", "Denmark", "Italy" , "Spain")),
                   Asia = data.frame(
                                  id = 1:4,
                                  oldValues = c("Baku", "Tokyo", "Kathmandu", "Singapore"),
                                  newValues = c("Azerbaijan", "Japan", "Nepal" , "Singapore")))
Countries
directory <- tempdir()
createRecodeDB(recodeListList = Countries,
               directory = directory,
               DBname = "Countries",
               fileType = "csv2")
createRecodeDB(recodeListList = Countries,
               directory = directory,
               DBname = "Countries",
               fileType = "xlsx")
# Get names of the recode lists in the data base
namesRecodeDB(directory = directory, DBname = "Countries", fileType = "csv2")
namesRecodeDB(directory = directory, DBname = "Countries", fileType = "xlsx")

}
