% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecodeDB.R
\name{namesRecodeDB}
\alias{namesRecodeDB}
\title{Get Names of Recode Data Base}
\usage{
namesRecodeDB(filePath)
}
\arguments{
\item{filePath}{Path to the \code{.xlsx} file in which the data base is stored.}
}
\description{
Get the names of the individual recode lists within a recode data base.
}
\examples{
# Create recode list data base using `createRecodeDB`
recodeListList <- list( Europe = data.frame(
                                  id = 1:4,
                                  oldValues = c("Berlin", "Copenhagen", "Rome", "Madrid"),
                                  newValues = c("Germany", "Denmark", "Italy" , "Spain")),
                        Asia = data.frame(
                                  id = 1:4,
                                  oldValues = c("Baku", "Tokyo", "Kathmandu", "Singapore"),
                                  newValues = c("Azerbaijan", "Japan", "Nepal" , "Singapore")))
recodeListList
filePath_temp <- tempfile(fileext = ".xlsx")
createRecodeDB(recodeListList = recodeListList, filePath = filePath_temp)
# Import data base
namesRecodeDB(filePath = filePath_temp)

}
