% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRecodeDB.R
\name{updateRecodeDB}
\alias{updateRecodeDB}
\title{Update Recode Data Base}
\usage{
updateRecodeDB(
  newRecodes,
  oldValues,
  newValues = "newValues",
  recodeDBPath,
  newRecodeDBPath,
  name,
  override = FALSE
)
}
\arguments{
\item{newRecodes}{A \code{data.frame} containing new recode information.}

\item{oldValues}{A character string of the column name containing the old values in the \verb{newRecodes data.frame}.}

\item{newValues}{A character string of the column name containing the newly recoded values in the \verb{newRecodes data.frame}.}

\item{recodeDBPath}{Path to the \code{.xlsx} file in which the data base is stored.}

\item{newRecodeDBPath}{Path to the \code{.xlsx} file in which the updated data base should be stored.}

\item{name}{Name of the specific recode list.}

\item{override}{Logical of length 1. Should existing recode pairs be overwritten when conflicting newer recode pairs
are present in the \code{newRecodes}?}
}
\description{
Update a recode data base stored in an Excel file.
Takes new recode information, compares it to the existing recode database, and updates the database accordingly.
}
\examples{
# example data base
oldDatabase <- list( Europe = data.frame(oldValues = c("Berlin", "Copenhagen", "Rome", "Madrid"),
                                         newValues = c("France", "Denmark", "Italy" , "Spain")),
                     Asia = data.frame(oldValues = c("Baku", "Tokyo", "Kathmandu", "Singapore"),
                                       newValues = c("Azerbaijan", "Japan", "Nepal" , "Singapore")))
oldDatabase
oldFilePath_temp <- tempfile(fileext = ".xlsx")
createRecodeDB(recodeListList = oldDatabase, filePath = oldFilePath_temp)
# new recode information
newRecodes <- data.frame( city = c("Berlin", "Paris", "Athens"),
                          country = c("Germany", "France", "Greece"))
newFilePath_temp <- tempfile(fileext = ".xlsx")
# update the data base without overwriting old information (the row containing "Berlin - France" keeps it's old value)
updateRecodeDB(newRecodes = newRecodes,
               oldValues = "city",
               newValues = "country",
               recodeDBPath = oldFilePath_temp,
               newRecodeDBPath = newFilePath_temp,
               name = "Europe",
               override = FALSE)
getRecodeDB(newFilePath_temp)
# update the data base, overwriting old information (the row containing "Berlin - France" get's updated)
updateRecodeDB(newRecodes = newRecodes,
               oldValues = "city",
               newValues = "country",
               recodeDBPath = oldFilePath_temp,
               newRecodeDBPath = newFilePath_temp,
               name = "Europe",
               override = TRUE)
getRecodeDB(newFilePath_temp)

}
