% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useRecodeList.R
\name{useRecodeList}
\alias{useRecodeList}
\title{Use Recode List}
\usage{
useRecodeList(df, varName, new_varName, recodeList)
}
\arguments{
\item{df}{A data.frame.}

\item{varName}{Column in \code{df} which should be recoded.}

\item{new_varName}{Column in which the recoded values will be stored.}

\item{recodeList}{A recode list: a data frame with the columns "oldValues" and "newValues".}
}
\value{
The \code{df} with a recoded \code{newValues}.
}
\description{
Add a new column to a data frame containing recoded values from a chosen column in that data frame.
}
\examples{
# example data frame
df <- data.frame(id = 1:4, country = c("Berlin", "Kairo", "Englant", "Schottland"))
# example recode list
recodeList <- data.frame(oldValues = c("Berlin", "Kairo", "Englant"),
                         newValues = c("Deutschland", "Ägypten", "England"))

useRecodeList(df, varName = "country", new_varName = "r_country", recodeList)

}
