% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useRecodeList.R
\name{useRecodeList}
\alias{useRecodeList}
\title{Use Recode List}
\usage{
useRecodeList(df, oldCol, recodeList)
}
\arguments{
\item{df}{A data.frame.}

\item{oldCol}{Column in \code{df} which should be recoded.}

\item{recodeList}{A recode list.}
}
\value{
The \code{df} with a recoded \code{newValues}.
}
\description{
Use Recode List.
}
\examples{
# tbd

}
